Separate createPages function from gatsby-node
Breakdown steps into separate functions
